More component refactoring
